remove file test, move mock data and update path
