+++ test/e2e/endpoints/category/emr.mock.data.ts
@@ -1,0 +1,59 @@
+import { EmrCategory } from '../../../../src/endpoints/category/emr/models/emr.entity';
+
+export const emrList: EmrCategory[] = [
+  {
+    id: 1,
+    category: 'Allergies and adverse drug reactions',
+    created_at: new Date(),
+  },
+  {
+    id: 2,
+    category: 'Chronic diseases',
+    created_at: new Date(),
+  },
+  {
+    id: 3,
+    category: 'Family medical history',
+    created_at: new Date(),
+  },
+  {
+    id: 4,
+    category: 'Illnesses and hospitalizations',
+    created_at: new Date(),
+  },
+  {
+    id: 5,
+    category: 'Laboratory test results',
+    created_at: new Date(),
+  },
+  {
+    id: 6,
+    category: 'Medications and dosing',
+    created_at: new Date(),
+  },
+  {
+    id: 7,
+    category: 'Prescription record',
+    created_at: new Date(),
+  },
+  {
+    id: 8,
+    category: 'Surgeries and other procedures',
+    created_at: new Date(),
+  },
+  {
+    id: 9,
+    category: 'Vaccinations',
+    created_at: new Date(),
+  },
+  {
+    id: 10,
+    category: 'Observations of daily living (ODLs)',
+    created_at: new Date(),
+  },
+  {
+    id: 11,
+    category: 'Others',
+    created_at: new Date(),
+  },
+];
