+++ test/e2e/endpoints/category/specialization.mock.data.ts
@@ -1,0 +1,64 @@
+import { SpecializationCategory } from '../../../../src/endpoints/category/specialization/models/specialization.entity';
+
+export const specializationList: SpecializationCategory[] = [
+  {
+    id: 1,
+    category: 'Assisted Reproductive Therapy (ART)/Infertility',
+    created_at: new Date(),
+  },
+  {
+    id: 2,
+    category: 'Cancer/Oncology',
+    created_at: new Date(),
+  },
+  {
+    id: 3,
+    category: 'Cardiovascular',
+    created_at: new Date(),
+  },
+  {
+    id: 4,
+    category: 'Cystic fbrosis',
+    created_at: new Date(),
+  },
+  {
+    id: 5,
+    category: 'Fetal intervention and therapy (FIT)',
+    created_at: new Date(),
+  },
+  {
+    id: 6,
+    category: 'Hematology',
+    created_at: new Date(),
+  },
+  {
+    id: 7,
+    category: 'Metabolic',
+    created_at: new Date(),
+  },
+  {
+    id: 8,
+    category: 'Neurology',
+    created_at: new Date(),
+  },
+  {
+    id: 9,
+    category: 'Pediatric and clinical',
+    created_at: new Date(),
+  },
+  {
+    id: 10,
+    category: 'Personalized medicine',
+    created_at: new Date(),
+  },
+  {
+    id: 11,
+    category: 'Prenatal',
+    created_at: new Date(),
+  },
+  {
+    id: 12,
+    category: 'Others',
+    created_at: new Date(),
+  },
+];
